Social-Distribution/socialdistribution/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.3
Django==5.1.2
django-encrypted-model-fields==0.6.5
django-model-utils==5.0.0
djangorestframework==3.15.2
Markdown==3.7
pillow==11.0.0
pycparser==2.22
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.7.0
